Folder organization, extract type
