+++ app/components/landing/LandingHero.tsx
