+++ app/components/lesson/LessonHero.tsx
