+++ app/components/lesson/LessonStep.tsx
