+++ app/components/lesson/SkillsLearned.tsx
